@@ -4,10 +4,12 @@ publish_to: "none"
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
+    sdk: flutter
+  flutter_localizations:
     sdk: flutter
   cached_network_image: ^2.5.1
   cupertino_icons: ^1.0.2
@@ -17,14 +19,16 @@ dependencies:
   flutter_facebook_login: ^3.0.0
   flutter_google_places: ^0.2.8
   flutter_svg: ^0.19.3
-  geolocation: ^1.1.2
+  #? geolocation: ^1.1.2
   get: ^3.26.0
   google_sign_in: ^5.0.1
   image_cropper: ^1.4.0
-  image_picker: #^0.7.3
+  image_picker: null
   intl: ^0.17.0
   permission_handler: ^6.1.1
+  qr_code_scanner: ^0.4.0-nullsafety.0
   shared_preferences: ^2.0.5
+  sign_in_with_apple: ^2.5.4
   url_launcher: ^6.0.2
 
 dependency_override:
