@@ -8,53 +8,53 @@ class SharedPreferenceHelper {
   static const LOCATION_KEY = 'location';
   static const FIRST_TIME_KEY = 'first-time';
 
-  static SharedPreferences preferences;
+  static SharedPreferences? preferences;
 
   static void storeAccessToken(String accessToken) {
-    preferences.setString(ACCESS_TOKEN_KEY, accessToken);
+    preferences?.setString(ACCESS_TOKEN_KEY, accessToken);
   }
 
-  static String get accessToken => preferences.getString(ACCESS_TOKEN_KEY);
+  static String? get accessToken => preferences?.getString(ACCESS_TOKEN_KEY);
 
   static void clear() {
-    preferences.clear();
+    preferences?.clear();
   }
 
   static void logout() {
-    preferences.remove(ACCESS_TOKEN_KEY);
-    preferences.remove(USER_KEY);
+    preferences?.remove(ACCESS_TOKEN_KEY);
+    preferences?.remove(USER_KEY);
   }
 
   static void storeLocation(List<double> coordinates) {
-    preferences.setString(LOCATION_KEY,
+    preferences?.setString(LOCATION_KEY,
         json.encode(List<dynamic>.from(coordinates.map((x) => x))));
   }
 
   static List<double> get location =>
-      preferences.getString(LOCATION_KEY)?.isEmpty ?? true
+      preferences?.getString(LOCATION_KEY)?.isEmpty ?? true
           ? []
-          : List<double>.from(json.decode(preferences.getString(LOCATION_KEY)))
+          : List<double>.from(json.decode(preferences?.getString(LOCATION_KEY)??''))
               .map((x) => x)
               .toList();
 
-  static void storeUser({UserResponse user, String response}) {
+  static void storeUser({UserResponse? user, String? response}) {
     if (user != null)
-      preferences.setString(USER_KEY, userResponseToJson(user));
+      preferences?.setString(USER_KEY, userResponseToJson(user));
     else {
       if (response == null || response.isEmpty)
         throw 'No value to store. Either a User object or a String response is required to store in preference.';
       else
-        preferences.setString(USER_KEY, response);
+        preferences?.setString(USER_KEY, response);
     }
   }
 
-  static UserResponse get user => preferences.getString(USER_KEY) == null
+  static UserResponse? get user => preferences?.getString(USER_KEY) == null
       ? null
-      : userResponseFromJson(preferences.getString(USER_KEY));
+      : userResponseFromJson(preferences?.getString(USER_KEY)??'');
 
   static void storeFirstTime() {
-    preferences.setBool(FIRST_TIME_KEY, false);
+    preferences?.setBool(FIRST_TIME_KEY, false);
   }
 
-  static void get isFirstTime => preferences.getBool(FIRST_TIME_KEY) ?? true;
+  static void get isFirstTime => preferences?.getBool(FIRST_TIME_KEY) ?? true;
 }
