@@ -4,11 +4,11 @@ part of 'app_button.dart';
 /// Created by Sunil Kumar from Boiler plate
 ///
 class AppCircleButton extends StatelessWidget {
-  final Color color;
-  final EdgeInsets padding;
-  final VoidCallback onPressed;
-  final Widget child;
-  final Size size;
+  final Color? color;
+  final EdgeInsets? padding;
+  final VoidCallback? onPressed;
+  final Widget? child;
+  final Size? size;
   AppCircleButton(
       {this.child, this.size, this.color, this.onPressed, this.padding});
   @override
